+++ src/Enums/SubscriptionStatus.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace LucaLongo\Subscriptions\Enums;
+
+enum SubscriptionStatus: string
+{
+    case INCOMPLETE = 'incomplete';
+    case INCOMPLETE_EXPIRED = 'incomplete_expired';
+    case TRIALING = 'trialing';
+    case ACTIVE = 'active';
+    case PAST_DUE = 'past_due';
+    case CANCELED = 'canceled';
+    case UNPAID = 'unpaid';
+    case PAUSED = 'paused';
+    case REVOKED = 'revoked';
+}