+++ src/Models/Subscription.php
@@ -2,6 +2,7 @@
 
 namespace LucaLongo\Subscriptions\Models;
 
+use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Casts\AsArrayObject;
 use Illuminate\Database\Eloquent\Casts\Attribute;
@@ -13,8 +14,13 @@ use Illuminate\Database\Eloquent\Relations\HasManyThrough;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Collection;
+use LucaLongo\Subscriptions\Actions\Subscriptions\CancelSubscription;
+use LucaLongo\Subscriptions\Actions\Subscriptions\RenewSubscription;
+use LucaLongo\Subscriptions\Actions\Subscriptions\RevokeSubscription;
+use LucaLongo\Subscriptions\Enums\SubscriptionStatus;
+use LucaLongo\Subscriptions\Models\Contracts\SubscriptionContract;
 
-class Subscription extends Model
+class Subscription extends Model implements SubscriptionContract
 {
     use HasUuids;
     use SoftDeletes;
@@ -29,15 +35,15 @@ class Subscription extends Model
     protected function casts(): array
     {
         return [
-            'starts_at' => 'datetime',
+            'auto_renew' => 'bool',
+            'status' => SubscriptionStatus::class,
             'ends_at' => 'datetime',
             'next_billing_at' => 'datetime',
             'price' => 'decimal:2',
-            'trial_starts_at' => 'datetime',
             'trial_ends_at' => 'datetime',
-            'grace_starts_at' => 'datetime',
             'grace_ends_at' => 'datetime',
             'revoked_at' => 'datetime',
+            'canceled_at' => 'datetime',
             'meta' => AsArrayObject::class,
         ];
     }
@@ -92,10 +98,6 @@ class Subscription extends Model
                 return false;
             }
 
-            if ($this->trial_starts_at && $this->trial_ends_at) {
-                return now()->isBetween($this->trial_starts_at, $this->trial_ends_at);
-            }
-
             if ($this->trial_ends_at) {
                 return now()->isBefore($this->trial_ends_at);
             }
@@ -109,10 +111,6 @@ class Subscription extends Model
         return Attribute::get(function () {
             if ($this->is_revoked) {
                 return false;
-            }
-
-            if ($this->grace_starts_at && $this->grace_ends_at) {
-                return now()->isBetween($this->grace_starts_at, $this->grace_ends_at);
             }
 
             if ($this->grace_ends_at) {
@@ -130,10 +128,6 @@ class Subscription extends Model
                 return false;
             }
 
-            if (now()->isBefore($this->starts_at)) {
-                return $this->is_trial_period;
-            }
-
             if ($this->ends_at && now()->isAfter($this->ends_at)) {
                 return $this->is_grace_period;
             }
@@ -149,16 +143,13 @@ class Subscription extends Model
             ->where(function (Builder $query) {
                 $query
                     ->where(fn (Builder $query) => $query
-                        ->where('starts_at', '<=', now())
-                        ->where(fn (Builder $query) => $query
-                            ->whereNull('ends_at')
-                            ->orWhere('ends_at', '>=', now())))
+                        ->whereNull('grace_ends_at')
+                        ->where('grace_ends_at', '>=', now())
+                    )
                     ->orWhere(fn (Builder $query) => $query
-                        ->where('trial_starts_at', '<=', now())
-                        ->where('trial_ends_at', '>=', now()))
-                    ->orWhere(fn (Builder $query) => $query
-                        ->where('grace_starts_at', '<=', now())
-                        ->where('grace_ends_at', '>=', now()));
+                        ->whereNull('ends_at')
+                        ->orWhere('ends_at', '>=', now())
+                    );
             })
         );
     }
@@ -176,5 +167,20 @@ class Subscription extends Model
     public function hasAllFeature(Collection $features): bool
     {
         return $this->features->pluck('code')->containsAll($features);
+    }
+
+    public function renew(?Carbon $endsAt = null): bool
+    {
+        return app(RenewSubscription::class)->execute($this, $endsAt);
+    }
+
+    public function cancel(?Carbon $endsAt = null): bool
+    {
+        return app(CancelSubscription::class)->execute($this, $endsAt);
+    }
+
+    public function revoke(?Carbon $revokesAt = null): bool
+    {
+        return app(RevokeSubscription::class)->execute($this, $revokesAt);
     }
 }