+++ CHANGELOG.md
@@ -2,29 +2,56 @@
 
 All notable changes to `laravel-subscriptions` will be documented in this file.
 
+## 1.0.0 - 2025-03-04
+
+### What's Changed
+
+* Add support to Laravel 10
+* Add support to payment gateways
+* Implements Stripe gateway
+* Add Stripe web-hooks
+* Bump dependabot/fetch-metadata from 2.2.0 to 2.3.0 by @dependabot in https://github.com/masterix21/laravel-subscriptions/pull/5
+* Bump aglipanci/laravel-pint-action from 2.4 to 2.5 by @dependabot in https://github.com/masterix21/laravel-subscriptions/pull/6
+* 1.0.0 by @masterix21 in https://github.com/masterix21/laravel-subscriptions/pull/7
+
+### New Contributors
+
+* @masterix21 made their first contribution in https://github.com/masterix21/laravel-subscriptions/pull/7
+
+**Full Changelog**: https://github.com/masterix21/laravel-subscriptions/compare/0.0.9...1.0.0
+
 ## 0.0.9 - 2024-07-25
+
 - [FIX] Subscriptions livewire component crashes when the invoice period is empty
 
 ## 0.0.8 - 2024-07-09
+
 - Bug fixing
 
 ## 0.0.7 - 2024-07-04
+
 - Add dateTime format to the Next Billing At column
 
 ## 0.0.6 - 2024-07-04
+
 - Add the next billing column to the subscriptions table
 
 ## 0.0.5 - 2024-07-03
+
 - Minor changes
 
 ## 0.0.4 - 2024-06-21
+
 - CreateSubscription now resolves the right model
 
 ## 0.0.3 - 2024-06-21
+
 - Improved Livewire component to manage a subscription
 
 ## 0.0.2 - 2024-06-10
+
 - Add provider references to subscriptions
 
 ## 0.0.1 - 2024-04-16
+
 Our start point (beta)